+++ src/game/Game.cpp
@@ -31,15 +31,15 @@ void Game::clearBackground(SDL_Renderer* renderTarget)
 void Game::drawDebugGrid(SDL_Renderer *renderTarget)
 {
     SDL_SetRenderDrawColor(renderTarget, 50, 50, 50, 255);
-    for (uint8_t x = 0; x < tileCountX; x++)
+    for (int x = 0; x < tileCountX; x++)
     {
-        uint16_t posX = x * tileSize + tileHalfSize;
+        int posX = x * tileSize + tileHalfSize;
         SDL_RenderDrawLine(renderTarget, posX, 0, posX, Application::renderer.viewport.w);
     }
 
-    for (uint8_t y = 0; y < tileCountY; y++)
+    for (int y = 0; y < tileCountY; y++)
     {
-        uint16_t posY = y * tileSize;
+        int posY = y * tileSize;
         SDL_RenderDrawLine(renderTarget, 0, posY, Application::renderer.viewport.h, posY);
     }
 }
